refactor: move repository classes in own file
